requirements.txt
fastapi>=0.115
uvicorn>=0.30
streamlit>=1.36
chromadb>=0.5
pandas>=2.2
python-dotenv>=1.0
langchain>=0.2
langchain-community>=0.2
langchain-openai>=0.1
pypdf>=4.2
tiktoken>=0.7
openai>=1.40